+++ includes/Core/class-vlt-plugin.php
@@ -29,6 +29,7 @@ class VLT_Plugin {
 		}
 
 		VLT_Settings::init();
+		VLT_CPT::init();
 		VLT_Admin::init();
 		VLT_Exporter::init();
 		VLT_REST_Controller::init();